add QHJ info update

+++ models/MultiGraCNNQHJ.py
@@ -24,7 +24,7 @@ class MultiGranularityCNNModel:
                                           shape=[None, self.config.X_maxlen, param.BaseConfig.word_dimension])
         self.input_X2 = tf.placeholder(name="inputX2_word", dtype=tf.float32,
                                           shape=[None, self.config.Y_maxlen, param.BaseConfig.word_dimension])
-        self.x2_label = tf.placeholder(name="inputX2_label", dtype=tf.float32,
+        self.x2_label = tf.placeholder(name="inputX2_label", dtype=tf.int32,
                                           shape=[None, self.config.Y_maxlen])
         self.y = tf.placeholder(name="target_y", dtype=tf.int32, shape=[None, 2])
